delete a recommendation

+++ service/static/js/rest_api.js
@@ -135,29 +135,34 @@ $(function () {
 
     });
 
-    // ****************************************
-    // Delete a Pet
+ // ****************************************
+    // Delete a Recommendation
     // ****************************************
 
     $("#delete-btn").click(function () {
 
-        let pet_id = $("#pet_id").val();
+        let recommendation_id = $("#recommendation_id").val();
+
+        if (!recommendation_id) {
+            flash_message("Please select a recommendation to delete.");
+            return;
+        }
 
         $("#flash_message").empty();
 
         let ajax = $.ajax({
             type: "DELETE",
-            url: `/pets/${pet_id}`,
+            url: `/api/recommendations/${recommendation_id}`,
             contentType: "application/json",
             data: '',
         })
 
-        ajax.done(function(res){
+        ajax.done(function (res) {
             clear_form_data()
-            flash_message("Pet has been Deleted!")
+            flash_message("Recommendation has been Deleted!")
         });
 
-        ajax.fail(function(res){
+        ajax.fail(function (res) {
             flash_message("Server error!")
         });
     });
